Initial commit of Excel import frontend module. Uses PHP Excel Reader (http://code.google.com/p/php-excel-reader/).
PHP Excel Reader classes are save in module/lib/spreadsheetExcelReader.class.php. Did minor modifications:
- renamed class and constructor for symfony autoload feature to work
- replaced deprecated split() function with explode() (php 5.3.0)
- suppressed failure notice for iconv() on line 1717 using @

This version is able to import documents from an excel. It saves the obligatory attributes name and adoption date. Call import/import to run script.
The Excel file path and filename are hardcoded for now and we use the example Excel file provided in downloads section. For this script to run the Excel is saved in web/uploads/. Start column and row for the import are also hardcoded, as is the amount of columns used for the document attributes. Later columns contain clause and tag information.

Checked sfPHPExcel, but so far no success. There two commented lines of creating the instance and loading the Excel. Please install sfPHPExcel plugin before testing the lines on your machine. I have not commited the plugin because I have not used it successfully yet.